+++ R/history.R
@@ -162,6 +162,26 @@ as.data.frame.keras_training_history <- function(x, ...) {
   df
 }
 
+to_keras_training_history <- function(history) {
+  
+  # turn history into an R object so it can be persited and
+  # and give it a class so we can write print/plot methods
+  params <- history$params
+  if (params$do_validation)
+    params$validation_samples <- dim(history$validation_data[[1]])[[1]]
+  
+  # normalize metrics
+  metrics <- history$history
+  metrics <- lapply(metrics, function(metric) {
+    as.numeric(lapply(metric, mean))
+  })
+  
+  # create history
+  keras_training_history(
+    params = params,
+    metrics = metrics
+  )
+}
 
 keras_training_history <- function(params, metrics) {